+++ src/main/java/org/example/payment/PaypalPay.java
@@ -1,8 +1,10 @@
 package org.example.payment;
 
+import java.math.BigDecimal;
+
 public class PaypalPay implements Payment {
     @Override
-    public boolean pay(float totalMoney) {
+    public boolean pay(BigDecimal totalMoney) {
         System.out.println("You should pay: " + totalMoney + " with Paypal");
         return true;
     }
